fix comment mode.
Closes #140.
If a line starts with ';', it is ignored but comment_mode is reset.
A ';' inside a line ignores just the portion following the ';' character.
The beginning of the line is still interpreted.
Probably a similar fix is necessary for the SDSUPPORT part; I cannot
test that part without an SD-shield.